// src/util/enums.ts
export enum SuitEnum {
    SPADES = 'spades',
    HEARTS = 'hearts',
    CLUBS = 'clubs',
    DIAMONDS = 'diamonds'
} 


export enum CardEnum {
    One = 'one',
    Two = 'two',
    Three = 'three',
    Four = 'four',
    Five = 'five',
    Six = 'six',
    Seven = 'seven',
    Eight = 'eight',
    Nine = 'nine',
    Ten = 'ten',
    Jack = 'jack',
    Queen = 'queen',
    King = 'king',
    Ace = 'ace'
}